Move item wheel menu code to its own file + Add Gate Keys to the menu
